Revert "rosy: wifi: Explicitly disable protected management frames"
This reverts the previous commit that explicitly disabled
Protected Management Frames (PMF) in the overlay.

The original fix was implemented to address compatibility issues
with WPA2/3 mixed access points, where the device would incorrectly
interact with PMF queries. However, after further review, it was found
that the `pmf=1` setting in the overlay was misunderstood, as it sets
PMF to be optional rather than explicitly enabling it.
This caused unnecessary disabling of PMF functionality for the device.

Reverting this change restores the correct behavior,
aligning with the intended configuration for WPA3 support,
without causing disruptions with PMF support.

This reverts commit 329c1a77d92d2b2dc2c281168336f4b7ef63c8ba.

Signed-off-by: Aditya Eka Pranata <aepranata@gmail.com>